GKE-engine/requirements.txt
Click==7.0
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Werkzeug==0.14.1
torch==1.6.0
sentence-transformers==0.3.2
spacy==2.3.2
nltk==3.5
laserembeddings
google-cloud-logging